tecknews/requirements.txt
# psycopg2-binary>=2.9,<3.0
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
beautifulsoup4==4.12.3
django-environ

# DjangoFilterBackend>=5.7.0
# celery
bs4==0.0.2
celery==5.4.0
html5lib==1.1
# rest_framework
requests==2.32.3
requests-html==0.10.0
lxml==5.2.2
lxml_html_clean==0.1.1
# Scrapy==2.11.2
# scrapy-djangoitem==1.1.1
flower==2.0.1













# amqp==5.2.0
# appdirs==1.4.4
# asgiref==3.8.1
# attrs==23.2.0
# Automat==22.10.0
# beautifulsoup4==4.12.3
# billiard==4.2.0
# bs4==0.0.2
# celery==5.4.0
# certifi==2024.7.4
# cffi==1.16.0
# charset-normalizer==3.3.2
# click==8.1.7
# click-didyoumean==0.3.1
# click-plugins==1.1.1
# click-repl==0.3.0
# colorama==0.4.6
# constantly==23.10.4
# cryptography==43.0.0
# cssselect==1.2.0
# defusedxml==0.7.1
# Django==5.0.7
# django-filter==24.2
# djangorestframework==3.15.2
# fake-useragent==1.5.1
# filelock==3.15.4
# h11==0.14.0
# html5lib==1.1
# hyperlink==21.0.0
# idna==3.7
# importlib_metadata==8.2.0
# incremental==22.10.0
# itemadapter==0.9.0
# itemloaders==1.3.1
# jmespath==1.0.1
# kombu==5.3.7
# lxml==5.2.2
# lxml_html_clean==0.1.1
# outcome==1.3.0.post0
# packaging==24.1
# parse==1.20.2
# parsel==1.9.1
# prompt_toolkit==3.0.47
# Protego==0.3.1
# pyasn1==0.6.0
# pyasn1_modules==0.4.0
# pycparser==2.22
# PyDispatcher==2.0.7
# pyee==11.1.0
# pyOpenSSL==24.2.1
# pyppeteer==2.0.0
# pyquery==2.0.0
# PySocks==1.7.1
# python-dateutil==2.9.0.post0
# queuelib==1.7.0
# requests==2.32.3
# requests-file==2.1.0
# requests-html==0.10.0
# Scrapy==2.11.2
# scrapy-djangoitem==1.1.1
# selenium==4.23.1
# service-identity==24.1.0
# six==1.16.0
# sniffio==1.3.1
# sortedcontainers==2.4.0
# soupsieve==2.5
# sqlparse==0.5.1
# tldextract==5.1.2
# tqdm==4.66.4
# trio==0.26.0
# trio-websocket==0.11.1
# Twisted==24.3.0
# twisted-iocpsupport==1.0.4
# typing_extensions==4.12.2
# tzdata==2024.1
# urllib3==1.26.19
# vine==5.1.0
# w3lib==2.2.1
# wcwidth==0.2.13
# webencodings==0.5.1
# websocket-client==1.8.0
# websockets==10.4
# wsproto==1.2.0
# zipp==3.19.2
# zope.interface==6.4.post2